+++ apps/analytics-web/CHANGELOG.md
@@ -1,5 +1,14 @@
 # analytics-web
 
+## 1.0.2
+
+## 1.0.2-alpha.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @codaco/analytics@9.0.0-alpha.0
+
 ## 1.0.1
 
 ### Patch Changes